% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouphmap.R
\name{ghmap}
\alias{ghmap}
\title{Grouphmap: Automated one-step common analysis of Batch expression profile}
\usage{
ghmap(Path, conRep, treRep, OrgDb, TYPE, UP = TRUE, cutree, TOP)
}
\arguments{
\item{Path}{"File storage path." Two groups are one file, and the control group should before the treatment group.}

\item{conRep}{Number of repetitions in control group.}

\item{treRep}{Number of repetitions in treatment group.}

\item{OrgDb}{org.Mm.eg.db, org.Hs.eg.db, and org.Sc.sgd.db. Please library().}

\item{TYPE}{"SYMBOL","ENSEMBOL"...}

\item{UP}{up is TRUE and down is FALSE}

\item{cutree}{heatmap can be devided multiple modules that make the functional difference and similarity of those group to be obvious}

\item{TOP}{the numeric.Such as 10 is the top 10 of GO analysis in each file.}
}
\value{
p1
}
\description{
Grouphmap: Automated one-step common analysis of Batch expression profile
}
\examples{
\donttest{library(org.Mm.eg.db)
Path<-system.file("extdata", package = "Grouphmap")
ghmap(Path,2,2,org.Mm.eg.db,"ENSEMBL",FALSE,2,10)
}
}
